refactor(ale_claw): Tier 4 — split context.py into token_estimation + compaction
3-way split of the 1323-line context.py:
- token_estimation.py (leaf, 44 LOC): SAFETY_MARGIN, FIXED_IMAGE_TOKENS, the
  base64 regex, estimate_message_tokens/estimate_messages_tokens — the shared
  primitives both the overflow core and compaction need. No harness deps.
- compaction.py (908 LOC): the budget-aware compaction + chunked summarization
  pipeline, CompactionResult, and repair_tool_use_result_pairing/
  ToolPairingRepairReport. Depends only on the leaf + external leaves
  (model_config/helper_runtime/canonical_sanitize) — never the context core.
- context.py (434 LOC): keeps truncation, is_context_overflow_error, and
  ContextOverflowCallback; re-exports both new modules so every
  `from .context import ...` (agent_loop/subagent_session/__init__/tests) is
  unchanged.
Core->compaction refs were docstring-only; SAFETY_MARGIN was the sole shared
runtime symbol (-> leaf), so context re-exporting compaction is cycle-free.
Dropped context's now-unused json/dataclass/_SKIP/call_helper_model imports.
Pure move; re-exports identical objects; full suite green (928 passed, 3 skipped).
